refactor(gui): cleanup trashcode and rename generic prompt module
