+++ shared/genericPrompt.ts
